Move conditional imports to utils module
