+++ xmltotabular/utils.py
@@ -1,6 +1,30 @@
+import sys
 from pathlib import Path
 
 from lxml import etree
+
+
+assert sys.version_info >= (3, 6), "Error: Python 3.6 or newer is required."
+
+if sys.version_info < (3, 7):
+    try:
+        from multiprocess import Pool, cpu_count
+    except ImportError:
+        sys.exit(
+            "Error: If running with Python < 3.7, the multiprocess library is required "
+            "(e.g. pip install multiprocess)."
+        )
+else:
+    from multiprocessing import Pool, cpu_count
+
+
+try:
+    from termcolor import colored
+except ImportError:
+
+    def colored(text, _color):
+        """ Dummy function in case termcolor is not available. """
+        return text
 
 
 def expand_paths(path_expr):
